requirements.txt
tensorflow==2.15.0
numpy
scikit-learn
pillow
gradio
